refactor!: remove `HdKeyring` as default export
